added first-time wizard and controller, to handle new database / upgrade database, etc.  See AMFirstTimeController and ui/AMFirstTimeWidget.

+++ source/acquaman/AMDacqScanController.cpp
@@ -142,7 +142,7 @@ void AMDacqScanController::play2d()
 		chRef.setNum(yy);
 		chRef.prepend("PGT_COUNTS");
 		AMChannel* ch = pScan_()->channel(chRef);
-		for(int xx=0; xx<pScan_()->count(); xx++) {
+		for(int xx=0; (unsigned)xx<pScan_()->count(); xx++) {
 			data2d->setZ(ch->value(xx), xx, yy);
 		}
 	}
