+++ source/acquaman/AMRegionsList.h
@@ -39,7 +39,7 @@ public slots:
 
 	/// Deletes the region refered to by index and renumbers subsequent regions accordingly. Returns true if successful, return false if index is invalid.
 	bool deleteRegion(size_t index){
-		if(index > count())
+		if((unsigned)index > (unsigned)count())
 			return false;
 		return regions_->removeRows(index, 1);
 	}
